ENYO-2128: Add MarqueeSupportAccessibilitySupport for preventing reading twice
https://jira2.lgsvl.com/browse/ENYO-2128
Enyo-DCO-1.1-Signed-off-by: Bongsub Kim <bongsub.kim@lgepartner.com>
