+++ lib/Item/ItemAccessibilitySupport.js
@@ -14,14 +14,6 @@ var
 * @mixin
 */
 module.exports = {
-
-	/**
-	* @private
-	*/
-	handlers: {
-		onRequestMarqueeStart  : 'marqueeStartHandler',
-		onSpotlightBlur : 'spotlightBlurdHandler'
-	},
 	
 	/**
 	* @private
@@ -32,25 +24,5 @@ module.exports = {
 			sup.apply(this, arguments);
 			this.setAttribute('aria-labelledby', !this.accessibilityLabel && enabled ? this.getId() : null);
 		};
-	}),
-
-	/**
-	* @private
-	*/
-	marqueeStartHandler: function (oSender, oEvent) {
-		var enabled = !this.accessibilityDisabled;
-		if (this._marquee_distance !== 0 && !this.$.marqueeText && enabled) {
-			this.setAttribute('aria-hidden', 'true');
-		}
-	},
-
-	/**
-	* @private
-	*/
-	spotlightBlurdHandler: function (oSender, oEvent) {
-		var enabled = !this.accessibilityDisabled;
-		if (this._marquee_distance !== 0 && enabled && (this.getAttribute('aria-hidden') === 'true')) {
-			this.setAttribute('aria-hidden', null);
-		}
-	}
+	})
 };
