+++ lib/Marquee/Marquee.js
@@ -12,11 +12,13 @@ var kind = require('enyo/kind'),
 	util = require('enyo/utils'),
 	Control = require('enyo/Control'),
 	Component = require('enyo/Component'),
-	Signals = require('enyo/Signals');
+	Signals = require('enyo/Signals'),
+	EnyoOptions = require('enyo/options');
 
 var
 	options = require('../options'),
-	HighlightText = require('../HighlightText');
+	HighlightText = require('../HighlightText'),
+	MarqueeSupportAccessibilitySupport = require('./MarqueeSupportAccessibilitySupport');
 
 var exports = module.exports = {};
 
@@ -190,6 +192,11 @@ var MarqueeSupport = {
 		onRequestMarqueeStart: '_marquee_stopPropagation',
 		onRequestMarqueeStop: '_marquee_stopPropagation'
 	},
+
+	/**
+	* @private
+	*/
+	mixins: EnyoOptions.accessibility ? [MarqueeSupportAccessibilitySupport] : null,
 
 	/**
 	* @private
